+++ src/main/java/com/leets/X/domain/user/domain/enums/LoginStatus.java
@@ -1,4 +1,4 @@
-package com.leets.X.domain.user.service;
+package com.leets.X.domain.user.domain.enums;
 
 public enum LoginStatus {
     LOGIN, REGISTER
